Quick fix for a issue with folders
Because of the changes made to around line 160 and the nice title it caused an issue where sometimes the files would sync old/new and would removed the new files causing the job to fail and the files to all be deleted

This is a temp fix only, and will need fully addressed at a later date

+++ arm/ripper/main.py
@@ -339,11 +339,14 @@ def main(logfile, job):
             logging.info("Permissions set successfully: " + str(perm_result))
 
         # remove empty directories
-        try:
-            os.rmdir(hboutpath)
-        except OSError:
-            logging.info(hboutpath + " directory is not empty.  Skipping removal. ")
-            pass
+        ## Same issue of removing files that have already been identified
+        ## TODO: fully fix this, this is only a temp fix
+        if hboutpath != final_directory:
+            try:
+                os.rmdir(hboutpath)
+            except OSError:
+                logging.info(hboutpath + " directory is not empty.  Skipping removal. ")
+                pass
 
         try:
             newpath
